backend/requirements.txt
requests
langchain
langgraph
langchain-google-genai
python-dotenv
flask
sentence-transformers
xgboost
numpy
flask-cors